server1/requirements.txt
flask

flask-sqlalchemy

flask-bcrypt

python-dotenv #(By using python-dotenv, you can separate your configuration details (like API keys, database URLs, etc.) from your code, providing a more secure and maintainable way to manage your application's configuration. It's particularly useful for keeping sensitive information out of your source code and making your application more portable.)

flask-session  #allows you to manage user sessions in your Flask application. It provides a way to store and retrieve data associated with a specific user across multiple requests, enhancing the functionality and personalization of your web application.

redis #Redis is a versatile and powerful tool that can enhance the performance and efficiency of applications, particularly those that require fast data access and manipulation.

flask-cors

scikit-learn

pandas
numpy
